refactor: 패키지 구조 변경
패키지 구조를 변경하였습니다. domain과 view로 나누고 domain에는 로직 관련 객체들이, view에는 인풋값을 받는 scanner와 값을 출력하는 print를 위치시켰습니다.
