+++ src/main/java/baseball/BaseballGame.java
@@ -1,11 +1,11 @@
 package baseball;
 
-import baseball.count.ball.BallCounter;
-import baseball.count.strike.StrikeCounter;
-import baseball.number.inputnumber.InputNumberParser;
-import baseball.scanner.NumberScanner;
-import baseball.print.MessagePrinter;
-import baseball.number.randomnumber.RandomNumberFactory;
+import baseball.domain.count.ball.BallCounter;
+import baseball.domain.count.strike.StrikeCounter;
+import baseball.domain.number.inputnumber.InputNumberParser;
+import baseball.view.scanner.NumberScanner;
+import baseball.view.print.MessagePrinter;
+import baseball.domain.number.randomnumber.RandomNumberFactory;
 
 import java.util.List;
