+++ src/main/java/baseball/domain/count/ball/BallResult.java
@@ -1,4 +1,4 @@
-package baseball.count.ball;
+package baseball.domain.count.ball;
 
 public class BallResult {
