+++ src/main/java/baseball/domain/count/strike/StrikeResult.java
@@ -1,4 +1,4 @@
-package baseball.count.strike;
+package baseball.domain.count.strike;
 
 public class StrikeResult {
