+++ src/main/java/baseball/view/print/Message.java
@@ -1,4 +1,4 @@
-package baseball.print;
+package baseball.view.print;
 
 public class Message {
