+++ src/test/java/baseball/number/inputnumber/InputNumberParserTest.java
@@ -1,5 +1,6 @@
 package baseball.number.inputnumber;
 
+import baseball.domain.number.inputnumber.InputNumberParser;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
