+++ src/test/java/baseball/scanner/NumberScannerTest.java
@@ -1,5 +1,6 @@
 package baseball.scanner;
 
+import baseball.view.scanner.NumberScanner;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
